Add authoirty

+++ src/hihapi/Controllers/Library/LibraryBookCategoriesController.cs
@@ -42,7 +42,7 @@ namespace hihapi.Controllers.Library
             }
 
             if (String.IsNullOrEmpty(usrName))
-                return Ok(_context.FinAccountCategories.Where(p => p.HomeID == null));
+                return Ok(_context.BookCategories.Where(p => p.HomeID == null));
 
             var rst0 = from ctgy in _context.BookCategories
                        where ctgy.HomeID == null
@@ -89,6 +89,32 @@ namespace hihapi.Controllers.Library
             {
                 HIHAPIUtility.HandleModalStateError(ModelState);
             }
+
+            // User
+            String usrName = String.Empty;
+            try
+            {
+                usrName = HIHAPIUtility.GetUserID(this);
+                if (String.IsNullOrEmpty(usrName))
+                {
+                    throw new UnauthorizedAccessException();
+                }
+            }
+            catch
+            {
+                throw new UnauthorizedAccessException();
+            }
+
+            // Check whether User assigned with specified Home ID
+            var hms = _context.HomeMembers.Where(p => p.HomeID == tbc.HomeID.Value && p.User == usrName).Count();
+            if (hms <= 0)
+            {
+                throw new UnauthorizedAccessException();
+            }
+
+            tbc.CreatedAt = DateTime.Now;
+            tbc.Createdby = usrName;
+
             _context.BookCategories.Add(tbc);
             await _context.SaveChangesAsync();
 
@@ -98,10 +124,32 @@ namespace hihapi.Controllers.Library
         [HttpDelete]
         public async Task<IActionResult> Delete([FromODataUri] int key)
         {
+            // User
+            String usrName = String.Empty;
+            try
+            {
+                usrName = HIHAPIUtility.GetUserID(this);
+                if (String.IsNullOrEmpty(usrName))
+                {
+                    throw new UnauthorizedAccessException();
+                }
+            }
+            catch
+            {
+                throw new UnauthorizedAccessException();
+            }
+
             var tbd = await _context.BookCategories.FindAsync(key);
             if (tbd == null)
             {
                 return NotFound();
+            }
+
+            // Check whether User assigned with specified Home ID
+            var hms = _context.HomeMembers.Where(p => p.HomeID == tbd.HomeID.Value && p.User == usrName).Count();
+            if (hms <= 0)
+            {
+                throw new UnauthorizedAccessException();
             }
 
             _context.BookCategories.Remove(tbd);
